+++ src/main/java/org/opentripplanner/routing/graph/Graph.java
@@ -11,6 +11,7 @@ import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.prefs.Preferences;
 import java.util.stream.Collectors;
+import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import javax.inject.Inject;
 import org.locationtech.jts.geom.Coordinate;
@@ -21,11 +22,11 @@ import org.opentripplanner.common.geometry.CompactElevationProfile;
 import org.opentripplanner.common.geometry.GraphUtils;
 import org.opentripplanner.ext.dataoverlay.configuration.DataOverlayParameterBindings;
 import org.opentripplanner.ext.geocoder.LuceneIndex;
-import org.opentripplanner.graph_builder.linking.VertexLinker;
 import org.opentripplanner.model.calendar.openinghours.OpeningHoursCalendarService;
 import org.opentripplanner.routing.edgetype.StreetEdge;
 import org.opentripplanner.routing.fares.FareService;
-import org.opentripplanner.routing.impl.StreetVertexIndex;
+import org.opentripplanner.routing.graph.index.StreetIndex;
+import org.opentripplanner.routing.linking.VertexLinker;
 import org.opentripplanner.routing.services.RealtimeVehiclePositionService;
 import org.opentripplanner.routing.services.notes.StreetNotesService;
 import org.opentripplanner.routing.vehicle_parking.VehicleParkingService;
@@ -59,7 +60,7 @@ public class Graph implements Serializable {
   @Nullable
   private final OpeningHoursCalendarService openingHoursCalendarService;
 
-  private transient StreetVertexIndex streetIndex;
+  private transient StreetIndex streetIndex;
 
   //Envelope of all OSM and transit vertices. Calculated during build time
   private WorldEnvelope envelope = null;
@@ -68,10 +69,6 @@ public class Graph implements Serializable {
 
   /* The preferences that were used for graph building. */
   public Preferences preferences = null;
-  // TODO OTP2: This is only enabled with static bike rental
-  public boolean hasBikeSharing = false;
-  public boolean hasParkRide = false;
-  public boolean hasBikeRide = false;
 
   /** True if OSM data was loaded into this Graph. */
   public boolean hasStreets = false;
@@ -305,7 +302,7 @@ public class Graph implements Serializable {
    */
   public void index(StopModel stopModel) {
     LOG.info("Index street model...");
-    streetIndex = new StreetVertexIndex(this, stopModel);
+    streetIndex = new StreetIndex(this, stopModel);
     LOG.info("Index street model complete.");
   }
 
@@ -318,7 +315,7 @@ public class Graph implements Serializable {
    * Get streetIndex, safe to use while routing, but do not use during graph build.
    * @see #getStreetIndexSafe(StopModel)
    */
-  public StreetVertexIndex getStreetIndex() {
+  public StreetIndex getStreetIndex() {
     return this.streetIndex;
   }
 
@@ -326,7 +323,7 @@ public class Graph implements Serializable {
    * Get streetIndex during graph build, both OSM street data and transit data must be loaded
    * before calling this.
    */
-  public StreetVertexIndex getStreetIndexSafe(StopModel stopModel) {
+  public StreetIndex getStreetIndexSafe(StopModel stopModel) {
     indexIfNotIndexed(stopModel);
     return this.streetIndex;
   }
@@ -432,10 +429,12 @@ public class Graph implements Serializable {
     return vehiclePositionService;
   }
 
+  @Nonnull
   public VehicleRentalService getVehicleRentalService() {
     return vehicleRentalStationService;
   }
 
+  @Nonnull
   public VehicleParkingService getVehicleParkingService() {
     return vehicleParkingService;
   }